requirements.txt
appdirs==1.4.4
build==1.0.3
buildozer==1.5.0
certifi==2023.11.17
charset-normalizer==3.3.2
colorama==0.4.6
Cython==3.0.6
distlib==0.3.8
docutils==0.20.1
filelock==3.13.1
idna==3.6
Jinja2==3.1.2
Kivy==2.2.1
Kivy-Garden==0.1.5
MarkupSafe==2.1.3
numpy==1.26.2
opencv-python==4.8.1.78
packaging==23.2
pexpect==4.9.0
Pillow==10.1.0
platformdirs==4.1.0
ptyprocess==0.7.0
Pygments==2.17.2
PyMySQL==1.1.0
pyproject_hooks==1.0.0
requests==2.31.0
sh==1.14.3
toml==0.10.2
tomli==2.0.1
urllib3==2.1.0
virtualenv==20.25.0
